+++ R/settings.R
@@ -37,7 +37,8 @@ Settings <- R6::R6Class("Settings",
             link_file_size_limit = 16L * 1024L^2,
             nline_to_break_succession = 2L,
             inlay_hints_minimum_arguments = 2L,
-            inlay_hints_minimum_argument_length = 2L
+            inlay_hints_minimum_argument_length = 2L,
+            nested_packages_depth = 0L
         )
     ),
     public = list(